requirements.txt
discord.py==2.4.0
asyncpg==0.29.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
PyNaCl==1.5.0
requests>=2.31.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
